Add static admin account
Now, all mail to postmaster and root goes to a special static admin
account.

This admin account is treated like a group account, but with a special
table in the database, just for itself.

This way, when the database fails, we can still receive mail for
postmaster and root.

This required quite some changes, due to this being the first static
account.