crontabs: Configure $MAILTO and $PATH
By default, cron sends mail to USER, where USER is the user executing
the cron job.

This is a problem with cron jobs not owned by root.

Instead, we configure all mail to be sent to root, which is generally
assumed to be able to receive mail (this is required for the crontabs
installed by some packages, for example).

We do not use the { mail_default_local_sender } Ansible variable to
avoid a hard dependency on the mail role.

The PATH default is also not deemed good.

Note that we have to make these configurations within each crontab.

See also crontab(5).