Backup
We create encrypted backups via restic and store them in S3.

To ensure basic consistency, we use LVM snapshots as basis for our
backups where possible.

Note that the SQL database (which we only read, but not manage) is not
backed up by us; this is deemed the task of the program managing the
database (lawandorga-backend-service).