webserver: Add HTTPS
This is in preparation for webmail.

This required some modularization.

The http-server role was renamed web/core/http and a new web/core/https
role was added that depends on the former and tls/acme.

Note that tls/acme in turn depends on web/core/http, so this could not
be merged with web/core/https, unless also merging tls/acme.